undo some unecessary abstraction
